:recycle: :truck: muevo dtos faltantes y refactor en atributos
- el precio de las entradas ahora es de tipo Float
- guardo en Hibernate los enums como STRING y no por posicion

+++ src/main/kotlin/ar/edu/unq/mientradita/model/Match.kt
@@ -12,7 +12,7 @@ class Match(
     @ManyToOne(fetch=FetchType.EAGER)
     val away: Team,
     val matchStartTime: LocalDateTime,
-    val ticketPrice: Double
+    val ticketPrice: Float
     ) {
 
     @Id
